Massive reorganization. Read the detailed description below:
Sorry. The way I was organizing folders was getting a little unwieldy and bloated from all the stuff I was downloading to test. From now on only the exact files that are being used in the game will go into the OfficialGame folder, organized the best we can. The Development folder is where we can keep all of our testing stuff. The only subfolder that will be uploaded to Github is the "TeamSharedAssets" folder, for now. You can safely make more folders next to it that won't be uploaded.